requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.38.0
gunicorn==20.1.0
ipykernel==6.21.2
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
# kiwisolver==1.4.4
# matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
# numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
# pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
# Pillow==9.4.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
pytz==2022.7.1
# pyzmq==25.0.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
starkbank-ecdsa==2.2.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
whitenoise==6.4.0
jupyter==1.0.0

